// frontend/core-ui/types/Organisation.ts
export type Status = 'rejected_by_origin' | 'rejected_by_relation' | 'approved'
export type OrganisationRole = 'participating' | 'funding' | 'hosting'
export type OrganisationSource = 'RSD' | 'ROR' | 'MANUAL'

// organisation colums used in editOrganisation.createOrganisation
// NOTE! update when type CoreOrganisationProps changes
export const columsForCreate = [
  'parent', 'slug', 'primary_maintainer', 'name', 'ror_id', 'is_tenant', 'website', 'logo_id',
]
export type CoreOrganisationProps = {
  parent: string | null
  slug: string | null
  primary_maintainer?: string | null
  name: string
  ror_id: string | null
  is_tenant: boolean
  website: string | null
  logo_id: string | null
}

// organisation colums used in editOrganisation.updateOrganisation
// NOTE! update when type Organisation changes
export const columsForUpdate = [
  'id',
  'description',
  ...columsForCreate
]
export type Organisation = CoreOrganisationProps & {
  id: string | null
  // about page content created by maintainer
  description: string | null
  parent_names?: string
  rsd_path?: string
}
// ...
export type OrganisationForOverview = Organisation & {
  id: string
  slug: string
  logo_id: string | null
  software_cnt: number | null
  project_cnt: number | null
  children_cnt: number | null
  release_cnt: number | null
  rsd_path: string
}

export type SoftwareOfOrganisation = {
  id: string
  slug: string
  brand_name: string
  short_statement: string
  is_published: boolean
  is_featured: boolean
  status: Status
  contributor_cnt: number | null
  mention_cnt: number | null
  updated_at: string
  organisation: string
}